[luci] Extract string util methods to Strings helper
This will relocate string related utility methods to helpers/Strings module.

ONE-DCO-1.0-Signed-off-by: SaeHie Park <saehie.park@gmail.com>
